+++ apps/docs/types.d.ts
@@ -2,8 +2,10 @@ import type { ComponentType, PropsWithChildren } from "react";
 import type { LiveCodeEditorProps } from "@/lib/liveCode/components/LiveCodeEditor/LiveCodeEditor";
 import type { DoAndDontTileProps } from "@/lib/mdx/components/DoAndDont/ExampleTile";
 
-interface DoAndDontProps
-  extends Omit<DoAndDontTileProps, "code" | "type" | "text"> {
+interface DoAndDontProps extends Omit<
+  DoAndDontTileProps,
+  "code" | "type" | "text"
+> {
   example?: string;
   exampleText?: string;
 }