+++ packages/components/src/components/Action/Action.browser.test.tsx
@@ -1,7 +1,7 @@
 import { render } from "vitest-browser-react";
 import { page, userEvent } from "vitest/browser";
 import { type FC } from "react";
-import Action from "@/components/Action";
+import Action, { type ActionProps } from "@/components/Action";
 import { Button, type ButtonProps } from "@/components/Button";
 import type { Mock } from "vitest";
 import Content from "@/components/Content/Content";
@@ -19,7 +19,13 @@ let asyncAction1: Mock;
 let asyncAction2: Mock;
 let actionHistory: string[];
 
+const unhandledRejectionHandler = vitest.fn();
+const unhandledErrorHandler = vitest.fn();
+
 beforeEach(() => {
+  window.addEventListener("unhandledrejection", unhandledRejectionHandler);
+  window.addEventListener("error", unhandledErrorHandler);
+
   vitest.useFakeTimers();
   vitest.resetAllMocks();
   actionHistory = [];
@@ -44,6 +50,8 @@ beforeEach(() => {
 afterEach(() => {
   vitest.runOnlyPendingTimers();
   vitest.useRealTimers();
+  window.removeEventListener("unhandledrejection", unhandledRejectionHandler);
+  window.removeEventListener("error", unhandledErrorHandler);
 });
 
 const TestButton: FC<ButtonProps> = (p) => (
@@ -271,15 +279,115 @@ describe("Confirmation modal", () => {
   });
 });
 
-describe("Feedback", () => {
-  test("is shown when sync action succeeds", async () => {
-    await render(
-      <Action onAction={syncAction1} showFeedback>
+describe("Global error handler", () => {
+  test("is called when sync action fails", async () => {
+    syncAction1.mockImplementation(() => {
+      throw new Error("Whoops");
+    });
+
+    const ui = () => (
+      <Action onAction={syncAction1}>
         <TestButton />
-      </Action>,
+      </Action>
     );
+
+    await render(ui());
     await clickTrigger();
+
+    expect(unhandledErrorHandler).toHaveBeenCalledWith(
+      expect.objectContaining({
+        error: expect.objectContaining({ message: "Whoops" }),
+      }),
+    );
+  });
+
+  test("is called when async action fails", async () => {
+    asyncAction1.mockImplementation(async () => {
+      await sleep();
+      throw new Error("Whoops");
+    });
+
+    const ui = () => (
+      <Action onAction={asyncAction1}>
+        <TestButton />
+      </Action>
+    );
+
+    const { rerender } = await render(ui());
+    await clickTrigger();
+    await vitest.advanceTimersByTimeAsync(asyncActionDuration * 2);
+    await rerender(ui());
+
+    expect(unhandledRejectionHandler).toHaveBeenCalledWith(
+      expect.objectContaining({
+        reason: expect.objectContaining({ message: "Whoops" }),
+      }),
+    );
+  });
+});
+
+describe("Feedback", () => {
+  const runTest = async (options: { props?: Partial<ActionProps> } = {}) => {
+    const ui = () => (
+      <Action
+        {...options.props}
+        onAction={
+          options.props?.onAction ??
+          (() => {
+            // empty sync action
+          })
+        }
+      >
+        <TestButton />
+      </Action>
+    );
+
+    const { rerender } = await render(ui());
+    await clickTrigger();
+    return {
+      advanceTime: (time = asyncActionDuration) =>
+        vitest.advanceTimersByTimeAsync(time),
+      rerender: () => rerender(ui()),
+    };
+  };
+
+  test("is shown when sync action succeeds", async () => {
+    await runTest({ props: { showFeedback: true } });
     expectIconInDom("check");
+  });
+
+  test("is shown when async action succeeds", async () => {
+    const { rerender, advanceTime } = await runTest({
+      props: { onAction: sleep, showFeedback: true },
+    });
+    await advanceTime();
+    await rerender();
+    expectIconInDom("check");
+  });
+
+  test("is shown when sync action fails", async () => {
+    await runTest({
+      props: {
+        onAction: () => {
+          throw new Error("Whoops");
+        },
+      },
+    });
+    expectIconInDom("x");
+  });
+
+  test("is shown when async action fails", async () => {
+    const { rerender, advanceTime } = await runTest({
+      props: {
+        onAction: async () => {
+          await sleep();
+          throw new Error("Whoops");
+        },
+      },
+    });
+    await advanceTime();
+    await rerender();
+    expectIconInDom("x");
   });
 
   test("is shown when set in props", async () => {
@@ -297,29 +405,12 @@ describe("Feedback", () => {
     expectIconInDom("x");
   });
 
-  test("is shown when sync action fails", async () => {
-    syncAction1.mockImplementation(() => {
-      throw new Error("Whoops");
-    });
-    await render(
-      <Action onAction={syncAction1} showFeedback>
-        <TestButton />
-      </Action>,
-    );
-    await clickTrigger();
-    expectIconInDom("x");
-  });
-
   test("is hidden after some time", async () => {
-    const ui = () => (
-      <Action onAction={syncAction1} showFeedback>
-        <TestButton />
-      </Action>
-    );
-    const { rerender } = await render(ui());
-    await clickTrigger();
-    await vitest.advanceTimersByTimeAsync(2000);
-    await rerender(ui());
+    const { rerender, advanceTime } = await runTest({
+      props: { showFeedback: true },
+    });
+    await advanceTime(2000);
+    await rerender();
     expectNoIconInDom();
   });
 });