+++ packages/remote-elements/src/auto-generated/RemoteTooltipTriggerElement.ts
@@ -16,6 +16,7 @@ export class RemoteTooltipTriggerElement extends FlowRemoteElement<RemoteTooltip
       delay: {},
       isDisabled: {},
       isOpen: {},
+      shouldCloseOnPress: {},
       trigger: {},
     };
   }