+++ packages/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0-alpha.648](https://github.com/mittwald/flow/compare/0.2.0-alpha.647...0.2.0-alpha.648) (2026-01-27)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
+# [0.2.0-alpha.647](https://github.com/mittwald/flow/compare/0.2.0-alpha.646...0.2.0-alpha.647) (2026-01-27)
+
+**Note:** Version bump only for package @mittwald/flow-react-components
+
 # [0.2.0-alpha.646](https://github.com/mittwald/flow/compare/0.2.0-alpha.645...0.2.0-alpha.646) (2026-01-27)
 
 **Note:** Version bump only for package @mittwald/flow-react-components